lucky/requirements.txt
PyMySQL==0.9.3
attrs==18.2.0
nose==1.3.7
